Extract parser logic from main.py
