+++ ethpm_cli/_utils/logger.py
@@ -1,0 +1,14 @@
+import logging
+import sys
+
+
+def setup_cli_logger() -> logging.Logger:
+    logger = logging.getLogger()
+    logger.setLevel(logging.INFO)
+    handler = logging.StreamHandler(sys.stdout)
+    handler.setLevel(logging.INFO)
+    logger.addHandler(handler)
+    return logger
+
+
+cli_logger = setup_cli_logger()
